+++ CHANGELOG.md
@@ -130,6 +130,11 @@ Changelog
   * GET `/1.1/endpoints/iax/<endpoint_iax_id>`
   * PUT `/1.1/endpoints/iax/<endpoint_iax_id>`
 
+* A new API for associating an endpoint iax with a trunk has been added:
+
+  * DELETE `/1.1/trunks/<trunk_id>/endpoints/iax/<endpoint_id>`
+  * PUT `/1.1/trunks/<trunk_id>/endpoints/iax/<endpoint_id>`
+
 
 17.17
 -----